pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cityscapes-object-removal"
version = "0.1.0"
description = "Object removal pipelines for street-view scenes"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch==2.9.1",
    "torchvision==0.24.1",
    "tqdm==4.67.1",
    "numpy",
    "pillow",
    "matplotlib",
    "pyyaml",
    "opencv-python",
    "diffusers",
    "transformers",
    "accelerate",
    "safetensors",
    "sentencepiece",
    "peft",
    "einops",
    "timm",
    "segment-anything",
]

[tool.setuptools.packages.find]
include = ["src*"]
